+++ tests/models/test_hooks.py
@@ -152,7 +152,7 @@ def test_prepare_batch_for_transfer():
     assert model.on_after_batch_transfer_hook_rank == 2
     assert batch_gpu.samples.device == batch_gpu.targets.device == expected_device
     assert torch.allclose(batch_gpu.samples.cpu(), torch.ones(5, 28))
-    assert torch.allclose(batch_gpu.targets.cpu(), torch.ones(5, 1) * 2)
+    assert torch.allclose(batch_gpu.targets.cpu(), torch.ones(5, 1, dtype=torch.long) * 2)
 
 
 @pytest.mark.parametrize(
